Persist user model to db

+++ forms/todo.go
@@ -1,6 +1,6 @@
 package forms
 
 type CreateTodoParams struct {
-	UserID uint64 `json:"user_id"`
+	UserID uint   `json:"user_id"`
 	Title  string `json:"title"`
 }
